pyproject.toml
[tool.poetry]
name = "webfinance"
version = "0.1.0"
description = "Website for finance ideas."
license = "MIT"
authors = ["Ivan Kyulev <ivandkyulev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
django-environ = "^0.11.2"
psycopg = "^3.1.18"
django-allauth = "^0.61.1"
google-auth = "^2.28.1"
google-cloud-secret-manager = "^2.18.2"
google-cloud-logging = "^3.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
